created HomeTabBarController class and restored the view controller to how it looked
